Update main.c to avoid warnings during build
Some changes to floats and doubles to avoid unnecessary build warnings

+++ v2.8.x-v2.7.0/lesson5/inter_less5_exer1_solution/src/main.c
@@ -372,7 +372,7 @@ int bme_read_sample(void)
 	int err;
 
 	int32_t datap = 0, datat = 0, datah = 0;
-	float pressure = 0.0, temperature = 0.0, humidity = 0.0;
+	float pressure = 0.0f, temperature = 0.0f, humidity = 0.0f;
 
 	/* STEP 9.1 - Store register addresses to do burst read */
 	uint8_t regs[] = {PRESSMSB, PRESSLSB, PRESSXLSB, \
@@ -407,14 +407,14 @@ int bme_read_sample(void)
 	bme280_compensate_humidity(&bmedata, datah);
 
 	/* Convert to proper format as per data sheet */
-	pressure 	= (float)bmedata.comp_press/256.0;
-	temperature = (float)bmedata.comp_temp/100.0;
-	humidity 	= (float)bmedata.comp_humidity/1024.0;
+	pressure 	= (float)bmedata.comp_press/256.0f;
+	temperature = (float)bmedata.comp_temp/100.0f;
+	humidity 	= (float)bmedata.comp_humidity/1024.0f;
 	
 	/* Print the uncompensated and compensated values */
-	LOG_INF("\tTemperature: \t uncomp = %d C \t comp = %8.2f C", datat, temperature);
-	LOG_INF("\tPressure:    \t uncomp = %d Pa \t comp = %8.2f Pa", datap, pressure);	
-	LOG_INF("\tHumidity:    \t uncomp = %d RH \t comp = %8.2f %%RH", datah, humidity);
+	LOG_INF("\tTemperature: \t uncomp = %d C \t comp = %8.2f C", datat, (double)temperature);
+	LOG_INF("\tPressure:    \t uncomp = %d Pa \t comp = %8.2f Pa", datap, (double)pressure);	
+	LOG_INF("\tHumidity:    \t uncomp = %d RH \t comp = %8.2f %%RH", datah, (double)humidity);
 
 	return 0;
 }
